+++ example/rbind.js
@@ -1,4 +1,4 @@
-var rbind = require('..').r;
+var rbind = require('..');
 var path = require('path');
 console.log(rbind(path, 'basename', '.css')('x.css'));
 // x