+++ test/xbind.js
@@ -38,21 +38,18 @@ test('xbind(ctx, f)', function (t) {
 });
 
 test('xbind(start, f)', function (t) {
-    t.same(xbind(arrayify)(2,3), []);
     t.same(xbind(1, arrayify)(2,3), [2]);
     t.same(xbind(2, arrayify)(2,3), [2,3]);
     t.end();
 });
 
 test('xbind(start, f, value)', function (t) {
-    t.same(xbind(arrayify, 1)(2,3), [1]);
     t.same(xbind(1, arrayify, 1)(2,3), [2,1]);
     t.same(xbind(2, arrayify, 1)(2,3), [2,3,1]);
     t.end();
 });
 
 test('xbind(start, f, value1, value2)', function (t) {
-    t.same(xbind(arrayify, 1, 0)(2,3), [1,0]);
     t.same(xbind(1, arrayify, 1, 0)(2,3), [2,1,0]);
     t.same(xbind(2, arrayify, 1, 0)(2,3), [2,3,1,0]);
     t.end();
@@ -60,7 +57,6 @@ test('xbind(start, f, value1, value2)', function (t) {
 
 test('xbind(start, ctx, "f", value1, value2)', function (t) {
     var o = { arrayify: arrayify };
-    t.same(xbind(o, 'arrayify', 1, 0)(2,3), [1,0]);
     t.same(xbind(1, o, 'arrayify', 1, 0)(2,3), [2,1,0]);
     t.same(xbind(2, o, 'arrayify', 1, 0)(2,3), [2,3,1,0]);
     t.end();